+++ R/format.R
@@ -14,6 +14,7 @@
 #' @param ... Additional arguments for methods.
 #'
 #' @export
+#' @seealso cli_vec
 #' @examples
 #' things <- c(rep("this", 3), "that")
 #' cli_format(things)
@@ -56,5 +57,34 @@ cli_format.character <- function(x, style = list(), ...) {
 cli_format.numeric <- function(x, style = list(), ...) {
   digits <- style$digits
   if (!is.null(digits)) x <- round(x, digits)
+  x
+}
+
+#' Add custom cli style to a vector
+#'
+#' @details
+#' You can this function to change the default parameters of
+#' [glue::glue_collapse()], see an example below.
+#'
+#' The style is added as an attribute, so operations that remove
+#' attributes will remove the style as well.
+#'
+#' @param x Vector, that will we collapsed by cli.
+#' @param style Style to apply to the vector. It is used as a theme on
+#' a `span` element that is created for the vector. You can set `vec_sep`
+#' and `vec_last` to modify the `sep` and `last` arguments of
+#' [glue::glue_collapse()]. See an example below.
+#'
+#' @export
+#' @seealso cli_format
+#' @examples
+#' v <- cli_vec(
+#'   c("foo", "bar", "foobar"),
+#'   style = list(vec_sep = " & ", vec_last = " & ")
+#' )
+#' cli_text("My list: {v}.")
+
+cli_vec <- function(x, style = list()) {
+  attr(x, "cli_style") <- style
   x
 }