+++ R/inline.R
@@ -3,18 +3,25 @@
 
 if (getRversion() >= "2.15.1") globalVariables("app")
 
-inline_generic <- function(app, class, x, style) {
-  xx <- paste0(style$before, x, style$after)
-  if (!is.null(style$fmt)) xx <- vcapply(xx, style$fmt)
+inline_generic <- function(app, x, style) {
+  vec_style <- attr(x, "cli_style")
+  before <- vec_style$before %||% style$before
+  after <- vec_style$after %||% style$after
+  fmt <- vec_style$fmt %||% style$fmt
+  xx <- paste0(before, x, after)
+  if (!is.null(fmt)) xx <- vcapply(xx, fmt)
   xx
 }
 
-inline_collapse <- function(x) {
-  if (length(x) >= 3) {
-    glue_collapse(x, sep = ", ", last = ", and ")
-  } else {
-    glue_collapse(x, sep = ", ", last = " and ")
-  }
+inline_collapse <- function(x, style = list()) {
+  vec_style <- attr(x, "cli_style")
+  sep <- vec_style$vec_sep %||%
+    style$vec_sep %||%
+    ", "
+  last <- vec_style$vec_last %||%
+    style$vec_last %||%
+    if (length(x) >= 3) ", and " else " and "
+  glue_collapse(x, sep = sep, last = last)
 }
 
 #' @importFrom glue glue glue_collapse
@@ -31,18 +38,25 @@ inline_transformer <- local({
     }, error = function(e) failed <<- TRUE ))
 
     if (!failed) {
+      if (inline_styling) return(res)
       rcls <- class(res)
       stls <- app$get_current_style()$`class-map`
       cls <- na.omit(match(rcls, names(stls)))[1]
-      if (!is.na(cls)) {
-        id <- clii__container_start(app, "span", class = stls[[cls]])
-        on.exit(clii__container_end(app, id), add = TRUE)
-        style_save <- style
-        on.exit(style <<- style_save, add = TRUE)
-        style <<- app$get_current_style()
-        res <- inline_generic(app, stls[[cls]], res, style)
+      if (is.na(cls)) class <- NULL else class <- stls[[cls]]
+      vec_style <- attr(res, "cli_style")
+      tid <- if (!is.null(vec_style)) {
+        app$add_theme(list(span = vec_style))
       }
-      if (inline_styling) return(res) else return(inline_collapse(res))
+      id <- clii__container_start(app, "span", class = class, theme = tid)
+      on.exit(clii__container_end(app, id), add = TRUE)
+      style_save <- style
+      on.exit(style <<- style_save, add = TRUE)
+      style <<- app$get_current_style()
+      res <- structure(
+        inline_generic(app, res, style),
+        cli_style = attr(res, "cli_style")
+      )
+      return(inline_collapse(res, style))
     }
 
     code <- glue_collapse(code, "\n")
@@ -78,7 +92,10 @@ inline_transformer <- local({
       .close = paste0(envir$marker, "}")
     )
 
-    inline_collapse(inline_generic(app, funname, out, style = style))
+    inline_collapse(
+      inline_generic(app, out, style = style),
+      style = style
+    )
   }
 })